Store kstitch CCA results in a Seurat object

Takes the list returned by link_shape_and_factors and writes results into the Seurat object in two places:

CSP and CEP scores are written into object-spanning DimReduc slots named <reduction_name>_csp and <reduction_name>_cep. Cells in the object that were not present in the CCA (e.g. cells excluded due to missing shape polygons) receive NA scores.

All other fit information is written to obj@misc$kstitch[[reduction_name]].

store_kstitch_results(
  obj,
  result,
  reduction_name = "kstitch",
  reduction_key_csp = "CSP_",
  reduction_key_cep = "CEP_"
)

Arguments

obj

A Seurat v5 object.

result

List as returned by link_shape_and_factors.

reduction_name

Character. Base name used to construct the DimReduc slot names (<reduction_name>_csp and <reduction_name>_cep) and the obj@misc$kstitch key. Must be unique across calls. Default "kstitch".

reduction_key_csp

Character. Key prefix for the CSP DimReduc. Default "CSP_".

reduction_key_cep

Character. Key prefix for the CEP DimReduc. Default "CEP_".

Value

The Seurat object with obj[["<reduction_name>_csp"]], obj[["<reduction_name>_cep"]], and obj@misc$kstitch[[reduction_name]] populated.

Details

To store multiple CCA results (e.g. cell vs nucleus, regressed vs unregressed, or different cell types), call this function once per result with a distinct reduction_name.

See also